Fifth_function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core==1.2.2
azure-storage-blob==12.2.0
azure-storage-file-datalake==12.0.0b7
